requirements.txt
ipykernel
pillow
pillow-heif
opencv-python
python-dotenv